feat(frontend): implement ConfessionForm component
